Fallback when intl is missing
This implements a mechanism to replace the locale aware part of the
function with differrent formatter classes.

For now it simply implements a fallback to a non-locale aware formatting
using DateTime->format() when the intl extension is missing. Future
extension could implement making use of strftime() if it still exist or
implement locale formatting in pure PHP.

For testing a environment variable STRFTIME_NO_INTL can be set to force
using the fallback.

Currently the test for the gregorian calendar jump fails, but I don't
understand the issue well enough to fix it.